added technical documentation
